updating verification with zod and opening tables

+++ src/controller/productsController.ts
@@ -25,6 +25,12 @@ class ProductsController {
 
       const { name, price } = bodySchema.parse(request.body)
 
+      const existProduct = await knex('products').where('name', name).first()
+
+      if (existProduct) {
+        return response.status(400).json({ message: 'Product already exists' })
+      }
+
       await knex<ProductsRepository>('products').insert({ name, price })
 
       return response
@@ -41,9 +47,29 @@ class ProductsController {
         name: z.string({ message: 'name is required' }).trim().min(6),
         price: z.number().gt(0, { message: 'price must be greater than 0' }),
       })
+
       const { name, price } = bodySchema.parse(request.body)
+
+      const product = await knex('products').where('id', request.params.id).first()
+
+      if (!product) {
+        return response.status(404).json({ message: 'Product not found' })
+      }
+      const existProduct = await knex('products')
+        .where('name', name)
+        .andWhereNot('id', request.params.id)
+        .first()
+
+      if (existProduct) {
+        return response.status(400).json({ message: 'Product name already exists' })
+      }
+
       await knex('products').where('id', request.params.id).update({ name, price })
-      return response.json({ message: 'Product updated successfully', product: { name, price } })
+
+      return response.json({
+        message: 'Product updated successfully',
+        product: { name, price },
+      })
     } catch (error) {
       next(error)
     }
@@ -51,7 +77,14 @@ class ProductsController {
 
   async delete(request: Request, response: Response, next: NextFunction) {
     try {
+      const product = await knex('products').where('id', request.params.id).first()
+
+      if (!product) {
+        return response.status(404).json({ message: 'Product not found' })
+      }
+
       await knex('products').where('id', request.params.id).delete()
+
       return response.json({ message: 'Product deleted successfully' })
     } catch (error) {
       next(error)
